I add css to my weekly specials page.
I linked it to the stylesheet and changed its background colours. I also created an aside with an image and extra information about the weekly special. I created a footer which sites the image in the aside. The weekly special timetable is in a div in the centre of the page.